Added in authorization to campgrounds and reviews

+++ Yelpcamp/routes/campgrounds.js
@@ -19,13 +19,14 @@ router.get("/new", isLoggedIn, (req,res)=>{
 router.post("/", isLoggedIn, catchAsync (async (req,res, next)=>{
     if(!req.body.campground) throw new ExpressErorr("Invalid Campground Data", 400);
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash('success', "Successfully created a new campground!")
     res.redirect(`/campgrounds/${campground._id}`)
 }))
 
 router.get("/:id", catchAsync (async (req, res, next)=>{
-    const campground = await Campground.findById(req.params.id).populate('reviews');
+    const campground = await Campground.findById(req.params.id).populate('reviews').populate("author");
     console.log(campground);
     if(!campground){
         req.flash('error', "Cannot find that campground!");
@@ -35,23 +36,38 @@ router.get("/:id", catchAsync (async (req, res, next)=>{
 }))
 
 router.get('/:id/edit', isLoggedIn, catchAsync(async (req,res, next)=>{
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
     if(!campground){
         req.flash('error', "Cannot find that campground!");
         res.redirect("/campgrounds")
+    }
+    if (!campground.author.equals(req.user._id)){
+        req.flash("error", "You do not have permission to do that!")
+        return res.redirect(`/campgrounds/${id}`)
     }
     res.render('campgrounds/edit', {campground});
 }))
 
 router.put('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)){
+        req.flash("error", "You do not have permission to do that!")
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'Successfully updated campground!');
     res.redirect(`/campgrounds/${campground._id}`)
 }));
 
 router.delete("/:id", isLoggedIn, catchAsync(async (req,res,next)=>{
-    const {id} = req.params;
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)){
+        req.flash("error", "You do not have permission to do that!")
+        return res.redirect(`/campgrounds/${id}`)
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', "Successfully deleted your review!")
     res.redirect('/campgrounds');
